+++ Jellyfin.Plugin.OpenDouban.Tests/Providers/OddbApiClientTest.cs
@@ -2,37 +2,36 @@ using System.Collections.Generic;
 using Microsoft.Extensions.DependencyInjection;
 using Xunit;
 using Xunit.Abstractions;
-using Jellyfin.Plugin.OpenDouban.Service;
 
 namespace Jellyfin.Plugin.OpenDouban.Tests
 {
-    public class ApiClientTest
+    public class OddbApiClientTest
     {
-        private readonly ApiClient apiClient;
+        private readonly OddbApiClient _oddbApiClient;
 
-        public ApiClientTest(ITestOutputHelper output)
+        public OddbApiClientTest(ITestOutputHelper output)
         {
-            var serviceProvider = ServiceUtils.BuildServiceProvider<ApiClient>(output);
-            this.apiClient = serviceProvider.GetService<ApiClient>();
+            var serviceProvider = ServiceUtils.BuildServiceProvider<OddbApiClient>(output);
+            _oddbApiClient = serviceProvider.GetService<OddbApiClient>();
         }
 
         [Fact]
         public void TestFullSearch() 
         {
-            List<ApiSubject> list = apiClient.FullSearch("Harry Potter").Result;
+            List<ApiSubject> list = _oddbApiClient.FullSearch("Harry Potter").Result;
             Assert.NotEmpty(list);
         }
 
         [Fact]
         public void TestPartialSearch() 
         {
-            List<ApiSubject> list = apiClient.PartialSearch("Harry Potter").Result;
+            List<ApiSubject> list = _oddbApiClient.PartialSearch("Harry Potter").Result;
             Assert.NotEmpty(list);
         }
 
         [Fact]
         public void TestGetBySid() {
-            ApiSubject result = apiClient.GetBySid("1295038").Result;
+            ApiSubject result = _oddbApiClient.GetBySid("1295038").Result;
             Assert.Equal("1295038", result.Sid);
             Assert.NotEmpty(result.Celebrities);
         }
